requirements.txt
Django>=1.9,<1.10
waitress>=0.8,<0.9
whitenoise>=3.2,<3.3

dj-database-url==0.3.0
django-debug-toolbar>=1.4,<1.5
django-sslify==0.2.7

python-decouple==3.0
pytz==2015.7
sqlparse>=0.1,<0.2

psycopg2>=2.6,<2.7
raven>=5.10,<5.11

asgi-redis==0.14.0
asgiref==0.14.0
autobahn==0.15.0
channels==0.17.0
contextlib2==0.5.1
daphne==0.14.2
defusedxml==0.4.1
django-extensions==1.6.7
django-rq==0.9.1
msgpack-python==0.4.7
oauthlib==1.1.2
PyJWT==1.4.1
python-isl==1.3
python3-openid==3.0.10
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.2
rq==0.6.0
six==1.10.0
Twisted==16.2.0
txaio==2.5.1
zope.interface==4.2.0
